+++ qcs-api-client-common/tests/token_refresh_test.rs
@@ -2,7 +2,9 @@
 
 use std::time::Duration;
 
-use qcs_api_client_common::configuration::{ClientConfiguration, ConfigSource};
+use qcs_api_client_common::configuration::{
+    ClientConfiguration, ConfigSource, SECRETS_READ_ONLY_VAR,
+};
 use time::{format_description::well_known::Rfc3339, OffsetDateTime};
 use toml_edit::DocumentMut;
 
@@ -34,6 +36,12 @@ async fn test_token_refresh() {
         secrets_path,
     } = configuration.source()
     {
+        if let Ok(ro_env) = std::env::var(SECRETS_READ_ONLY_VAR) {
+            if matches!(ro_env.to_lowercase().as_str(), "true" | "yes" | "1") {
+                // In this case, the file will *not* be updated.
+                return;
+            }
+        }
         let toml = std::fs::read_to_string(secrets_path)
             .unwrap()
             .parse::<DocumentMut>()